for语句解析(迭代中)
